refactor(entities): stock 엔티티 구조 리팩토링
- 기존 api, hooks, repository 디렉토리 제거
- stock.model.ts, stock.type.ts로 모델 구조 단순화
- 엔티티 레이어에서 비즈니스 로직 분리
