--- apps/web/src/entities/stock/api/index.ts
@@ -1,8 +1,0 @@
-export {
-    getHotStock,
-    getOverseasStockPrice,
-    getSelectedStocks,
-    getStocks,
-    getStockBySymbol,
-    selectStock,
-} from './stock';
