--- apps/web/src/entities/stock/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useSelectedStocks } from './stock';
