+++ apps/web/src/entities/stock/index.ts
@@ -1,28 +1,2 @@
-// API Functions - Client Side
-export { getStocks, getSelectedStocks, getOverseasStockPrice, getHotStock, getStockBySymbol, selectStock } from './api';
-
-// Hooks
-export { useSelectedStocks } from './hooks';
-
-// Additional hooks from hooks/stock.ts
-export {
-    useStocks,
-    useInfiniteSelectedStocks,
-    useOverseasStockPrice,
-    useSuspenseOverseasStockPrice,
-    useSuspenseHotStocks,
-    useStockBySymbol,
-} from './hooks/stock';
-
-// Types
-export type {
-    // Server Types
-    MarketCode,
-    DbStock,
-    KISAccessTokenResponse,
-    OverseasStockConditionSearchResponse,
-    OverseasStockPriceResponse,
-    OverseasStockInfoResponse,
-    // Client Types
-    Stock,
-} from './model';
+export type { ApiStockData, DbStock, MarketCode, Stock } from './model';
+export { StockModel } from './model';
