+++ apps/web/src/entities/stock/model/index.ts
@@ -1,10 +1,3 @@
-export type {
-    KISAccessTokenResponse,
-    OverseasStockConditionSearchResponse,
-    MarketCode,
-    OverseasStockPriceResponse,
-    OverseasStockInfoResponse,
-    DbStock,
-} from './types.server';
+export type { Stock, DbStock, ApiStockData, MarketCode } from './stock.type';
 
-export type { Stock } from './types';
+export { StockModel } from './stock.model';
