--- apps/web/src/entities/stock/repository/index.ts
@@ -1,1 +1,0 @@
-export { StockRepository } from './stock.server';
